requirements.txt
# Core dependencies
snowflake-snowpark-python[pandas]>=1.11.0
typer>=0.9.0
rich>=13.0.0
PyYAML>=6.0.0
pydantic>=2.0.0
psutil>=5.9.0

# Database drivers
# SQL Server
pyodbc>=5.0.0

# PostgreSQL
psycopg2-binary>=2.9.0

# MySQL
PyMySQL>=1.1.0

# Oracle
oracledb>=1.4.0

# Databricks
databricks-sql-connector>=3.0.0

# Notebook dependencies (optional for interactive development)
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0

# Development dependencies (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
mypy>=1.7.0 